feat: moved settings drawer to application level
