+++ apps/publisher/src/components/publisher-video-view/video-settings-drawer/utils.ts
@@ -2,7 +2,6 @@ import { VideoCodec } from '@millicast/sdk';
 
 import { Resolution } from '@millicast-react/use-multi-media-devices';
 import { Bitrate } from '@millicast-react/use-publisher';
-import { SimulcastQuality } from '@millicast-react/use-viewer';
 
 export const bitrateElementResolver = (element: unknown) => {
   const bitrateElement = element as Bitrate;
@@ -18,24 +17,6 @@ export const codecElementResolver = (element: unknown) => ({
   id: element as string,
   label: element as string,
 });
-
-export const deviceElementResolver = (element: unknown) => {
-  const deviceElement = element as InputDeviceInfo;
-  return {
-    data: deviceElement,
-    id: deviceElement.deviceId,
-    label: deviceElement.label,
-  };
-};
-
-export const qualityElementResolver = (element: unknown) => {
-  const quality = element as SimulcastQuality;
-  return {
-    data: quality,
-    id: quality.streamQuality,
-    label: quality.streamQuality,
-  };
-};
 
 export const resolutionElementResolver = (element: unknown) => {
   const resolution = element as Resolution;
