+++ apps/viewer/src/components/viewer-video-view/types.ts
@@ -1,0 +1,9 @@
+import { VideoViewProps } from '@millicast-react/video-view';
+
+import { VideoSettingsDrawerProps } from './video-settings-drawer';
+
+export interface ViewerVideoViewProps {
+  disableSettings?: boolean;
+  settingsProps: Omit<VideoSettingsDrawerProps, 'isOpen' | 'onClose'>;
+  videoProps: VideoViewProps;
+}
