+++ apps/viewer/src/components/viewer-video-view/video-settings-drawer/utils.ts
@@ -1,0 +1,10 @@
+import { SimulcastQuality } from '@millicast-react/use-viewer';
+
+export const qualityElementResolver = (element: unknown) => {
+  const quality = element as SimulcastQuality;
+  return {
+    data: quality,
+    id: quality.streamQuality,
+    label: quality.streamQuality,
+  };
+};
